Add add_edges, add_elements, add_element_attributes
to ElementService and HierarchyService.
Add hash method to TM1Object base class.
